+++ packages/bezier-react/src/components/Toast/ToastProvider.tsx
@@ -1,25 +1,20 @@
 /* External dependencies */
-import React, { useCallback, useEffect, useMemo, useState } from 'react'
+import React, { useCallback } from 'react'
 import { createPortal } from 'react-dom'
-import { noop } from 'lodash-es'
 
 /* Internal dependencies */
 import { css, TransitionDuration } from 'Foundation'
 import { getRootElement } from 'Utils/domUtils'
 import ToastContext from './ToastContext'
 import {
-  OnDismissCallback,
-  defaultOptions,
-  ToastOptions,
   ToastPlacement,
-  ToastId,
   ToastProviderProps,
   ToastType,
 } from './Toast.types'
 import ToastContainer from './ToastContainer'
 import ToastController from './ToastController'
 import ToastElement from './ToastElement'
-import ToastService from './ToastService'
+import useToastProviderValues from './useToastContextValues'
 
 const { Provider } = ToastContext
 
@@ -27,74 +22,12 @@ function ToastProvider({
   autoDismissTimeout = 3000,
   children = [],
 }: ToastProviderProps) {
-  const leftToastService = useMemo(() => new ToastService(), [])
-  const rightToastService = useMemo(() => new ToastService(), [])
-  const [leftToasts, setLeftToasts] = useState<ToastType[]>([])
-  const [rightToasts, setRightToasts] = useState<ToastType[]>([])
-
-  const add = useCallback((content: string, options: ToastOptions = defaultOptions) => {
-    let result = ''
-    if (options.rightSide) {
-      result = rightToastService.add(content, options)
-      setRightToasts(rightToastService.getToasts())
-    } else {
-      result = leftToastService.add(content, options)
-      setLeftToasts(leftToastService.getToasts())
-    }
-    return result
-  }, [
-    leftToastService,
-    rightToastService,
-  ])
-
-  const remove = useCallback((id: ToastId) => {
-    if (leftToastService.remove(id)) {
-      setLeftToasts(leftToastService.getToasts())
-    }
-    if (rightToastService.remove(id)) {
-      setRightToasts(rightToastService.getToasts())
-    }
-  }, [
-    leftToastService,
-    rightToastService,
-  ])
-
-  const removeAll = useCallback(() => {
-    leftToastService.removeAll()
-    rightToastService.removeAll()
-    setLeftToasts(leftToastService.getToasts())
-    setRightToasts(rightToastService.getToasts())
-  }, [
-    leftToastService,
-    rightToastService,
-  ])
-
-  const handleDismiss = useCallback((id: ToastId, callback: OnDismissCallback = noop) => {
-    callback(id)
-    remove(id)
-  }, [remove])
-
-  useEffect(() => {
-    setLeftToasts(leftToastService.getToasts())
-    setRightToasts(rightToastService.getToasts())
-  }, [
-    leftToastService,
-    rightToastService,
-  ])
-
-  const ToastContextValue = useMemo(() => ({
-    add,
-    remove,
-    removeAll,
+  const toastContextValue = useToastProviderValues()
+  const {
     leftToasts,
     rightToasts,
-  }), [
-    add,
-    remove,
-    removeAll,
-    leftToasts,
-    rightToasts,
-  ])
+    dismiss,
+  } = toastContextValue
 
   const createContainer = useCallback((placement: ToastPlacement, toasts: ToastType[]) => (
     <ToastContainer
@@ -112,6 +45,7 @@ function ToastProvider({
         id,
         onDismiss,
         zIndex,
+        version,
       }) => (
         <ToastController
           key={id}
@@ -126,19 +60,20 @@ function ToastProvider({
           content={content}
           iconName={iconName}
           component={ToastElement}
-          onDismiss={() => handleDismiss(id, onDismiss)}
+          onDismiss={() => dismiss(id, onDismiss)}
           transform={css``}
           zIndex={zIndex}
+          version={version}
         />
       )) }
     </ToastContainer>
   ), [
     autoDismissTimeout,
-    handleDismiss,
+    dismiss,
   ])
 
   return (
-    <Provider value={ToastContextValue}>
+    <Provider value={toastContextValue}>
       { children }
       { createPortal(
         [